Refactored Gpio to support different libraries
