+++ .idea/QtSettings.xml
@@ -2,5 +2,25 @@
 <project version="4">
   <component name="QtSettings">
     <option name="myCurrentProfile" value="Debug" />
+    <option name="mySettingsPerProfile">
+      <map>
+        <entry key="Debug">
+          <value>
+            <PerProfileState>
+              <option name="myQmlPath" value="/usr/lib/x86_64-linux-gnu/qt6/qml" />
+              <option name="myQtBinPath" value="/usr/lib/qt6/bin" />
+            </PerProfileState>
+          </value>
+        </entry>
+        <entry key="Release">
+          <value>
+            <PerProfileState>
+              <option name="myQmlPath" value="/usr/lib/x86_64-linux-gnu/qt6/qml" />
+              <option name="myQtBinPath" value="/usr/lib/qt6/bin" />
+            </PerProfileState>
+          </value>
+        </entry>
+      </map>
+    </option>
   </component>
 </project>
