+++ config/DigitalInputConfig.h
@@ -12,12 +12,15 @@ class DigitalInputConfig : public ConfigBase
 public:
   static constexpr auto type = "digitalinput";
 
-  DigitalInputConfig() : ConfigBase(type) {}
+  DigitalInputConfig() : ConfigBase(type), m_debounce(true) {}
   explicit DigitalInputConfig(const char * subtype) : ConfigBase(subtype) {}
   ~DigitalInputConfig() override  = default;
 
   void initialise(const nlohmann::json& json) override
   {
+    if (json.contains("debounce")) {
+      m_debounce = json["debounce"];
+    }
     if (json.contains("lines")) {
       for (auto& line : json["lines"]) {
         m_lines.push_back(line);
@@ -39,5 +42,6 @@ public:
 protected:
   std::vector<uint32_t> m_lines;
   std::string m_settingPath;
+  bool m_debounce;
 };
 #endif //CUTESDR_VK6HL_DIGITALINPUTCONFIG_H
