+++ control/ControlSinkFactory.cpp
