+++ control/device/gpio/GpioException.cpp
