+++ control/device/gpio/GpioException.h
