+++ control/device/gpio/digital/GpioRotaryEncoder.cpp
