+++ control/device/gpio/digital/GpioRotaryEncoder.h
