+++ control/device/gpio/impl/gpiod/GpioLinesImplGpiod.cpp
@@ -1,0 +1,172 @@
+//
+// Created by murray on 18/9/25.
+//
+
+#include "GpioLinesImplGpiod.h"
+#include "../../GpioException.h"
+#include <qdebug.h>
+
+GpioLinesImplGpiod::GpioLinesImplGpiod(gpiod_line_request * pLineRequest, const char* consumer) :
+  m_pCallback(nullptr),
+  m_pLineRequest(pLineRequest),
+  m_consumer(consumer)
+{
+  m_pEventBuffer = gpiod_edge_event_buffer_new(64);
+}
+
+GpioLinesImplGpiod::~GpioLinesImplGpiod()
+{
+  gpiod_edge_event_buffer_free(m_pEventBuffer);
+}
+
+void
+GpioLinesImplGpiod::startCallbacks(Callback* callback)
+{
+  if (m_pCallback != nullptr) {
+    throw GpioException("Line state callback already set");
+  }
+  // std::lock_guard<std::mutex> lock(m_callbackMutex);
+  m_callbackMutex.lock();
+  m_pCallback = callback;
+  m_callbackMutex.unlock();
+  QThread::start();
+}
+
+void
+GpioLinesImplGpiod::stopCallbacks()
+{
+  m_callbackMutex.lock();
+  m_pCallback = nullptr;
+  m_callbackMutex.unlock();
+  wait();
+}
+
+
+void
+GpioLinesImplGpiod::release()
+{
+  if (m_pLineRequest) {
+    gpiod_line_request_release(m_pLineRequest);
+    m_pLineRequest = nullptr;
+  }
+}
+
+int
+GpioLinesImplGpiod::getLineValue(uint32_t line)
+{
+  return gpiod_line_request_get_value(m_pLineRequest, line);
+}
+
+int
+GpioLinesImplGpiod::waitEdgeEvents(int64_t timeout_ns) const
+{
+  return gpiod_line_request_wait_edge_events(m_pLineRequest, timeout_ns);
+}
+
+int
+GpioLinesImplGpiod::readEdgeEvents(struct gpiod_edge_event_buffer* buf, size_t max_events) const
+{
+  return gpiod_line_request_read_edge_events(m_pLineRequest, buf, max_events);
+}
+
+void
+GpioLinesImplGpiod::run()
+{
+  // readInitialInputStates();
+  // gpiod_edge_event_buffer* eventBuff = gpiod_edge_event_buffer_new(64);
+  // if (!eventBuff) {
+  //   return; //TODO: flag the error
+  // }
+
+  LineStateMap debouncedLines;
+
+  constexpr int64_t idleTimeout = 100'000'000;
+  bool haveCallback = m_pCallback != nullptr;
+  while (haveCallback) {
+    // int numChanges = m_lines.debounce(changes);
+    // for (const auto& change : changes) {
+    //   qDebug() << "Line " << change.line << " changed to " << (int)change.value;
+    // }
+    int wr = waitEdgeEvents(idleTimeout);
+    if (wr < 0) {
+      // error; consider logging and continuing or breaking
+      continue;
+    }
+    if (wr == 0) {
+      // timeout: do periodic tasks if needed
+      continue;
+    }
+    qDebug() << "-----------------------";
+    debouncedLines.clear();
+    int numChanges = debounce(debouncedLines);
+    if (numChanges > 0) {
+      std::lock_guard<std::mutex> lock(m_callbackMutex);
+      if (m_pCallback) {
+        m_pCallback->callback(debouncedLines);
+      } else {
+        haveCallback = false;
+      }
+    }
+    // if (numHandled == 0 && numChanges > 0) {
+    //   qDebug() << "Unhandled changes:";
+    //   for (auto& infoPair : debouncedLines) {
+    //     GpioLines::LineState& info = infoPair.second;
+    //     qDebug() << "Line: " << info.line << ", LastRisingTime: " << info.lastRisingTime << ", LastFallingTime: " << info.lastFallingTime << ", Value: " << (int)info.value;
+    //   }
+    //
+    // }
+  }
+}
+
+int
+GpioLinesImplGpiod::debounce(LineStateMap& changes)
+{
+  constexpr int64_t debounceTimeout= 200'000'000;
+  //std::unordered_map<uint32_t, DebounceInfo> m_debouncedLines;
+  int waitResult = waitEdgeEvents(debounceTimeout);
+
+  if (waitResult < 0) {
+    throw GpioException("Error waiting for edge events");
+  }
+  // Whizz through the event storm until (we think) it's over
+  while(waitResult != 0) {
+    int numEvents = readEdgeEvents(m_pEventBuffer, 64);
+
+    for (int i = 0; i < numEvents; ++i) {
+
+      gpiod_edge_event* ev = gpiod_edge_event_buffer_get_event(m_pEventBuffer, i);
+      uint32_t line = gpiod_edge_event_get_line_offset(ev);
+      gpiod_edge_event_type type = gpiod_edge_event_get_event_type(ev);
+      uint64_t timestamp = gpiod_edge_event_get_timestamp_ns(ev);
+
+      auto stateIter = m_lineStates.find(line);
+      if (stateIter == m_lineStates.end()) {
+        throw GpioException("Received event for unknown line");
+      }
+      LineState& info = stateIter->second;
+      info.changed = true;
+      if (type == GPIOD_EDGE_EVENT_RISING_EDGE) {
+        info.lastRisingTime = timestamp;
+        info.value = 1;
+      } else {
+        info.lastFallingTime = timestamp;
+        info.value = 0;
+      }
+      changes[line] = info;
+    }
+    waitResult = waitEdgeEvents(debounceTimeout);
+    if (waitResult < 0) {
+      throw GpioException("Error waiting for edge events");
+    }
+  }
+  // Now read the affected line values to make sure we have that part of the story straight.
+  for (auto& pair : m_lineStates) {
+    LineState& info = pair.second;
+    if (info.changed) {
+      info.value = static_cast<uint8_t>(getLineValue(pair.first));
+      changes[pair.first] = info;
+      info.changed = false;
+    }
+  }
+  return changes.size();
+}
