+++ control/device/usb/UsbException.cpp
